move uflow boundary class to features submodule
